chore: rename shopping list files
