+++ src/pages/shopping-list/shopping-list-select-mode.tsx
@@ -1,7 +1,7 @@
 import { type Signal } from '@preact/signals';
 import { type JSX } from 'preact';
 import { type Item } from '@/responses';
-import { ShoppingItemSelection } from './shopping-item-selection';
+import { SelectShoppingItem } from './select-shopping-item';
 import { ShoppingListWrapper } from './shopping-list-wrapper';
 
 type ShoppingItemsProps = {
@@ -9,10 +9,10 @@ type ShoppingItemsProps = {
   invalidate: () => void;
 };
 
-export function ShoppingListSelection({ items, invalidate }: ShoppingItemsProps): JSX.Element {
+export function ShoppingListSelectMode({ items, invalidate }: ShoppingItemsProps): JSX.Element {
   return (
     <ShoppingListWrapper items={items}>
-      <>{items.value?.map((item) => <ShoppingItemSelection item={item} key={item.id} invalidateItems={invalidate} />)}</>
+      <>{items.value?.map((item) => <SelectShoppingItem item={item} key={item.id} invalidateItems={invalidate} />)}</>
     </ShoppingListWrapper>
   );
 }
